cleanFolder() remove all unknow file. Compare full array in FolderTest. cleanFolder() don't remove file that we now & that we don't whant but remove all exept usefull file.

+++ src/main/java/fr/formiko/usuel/images/image.java
@@ -214,10 +214,12 @@ public class image{
    *@version 1.3
    */
   public static void clearTemporaire(){
-    File docier = new File(REP_TEMPORARY);
-    File[] fichiers = docier.listFiles();
-    for (File f : fichiers ) {
-      if(isImage(f)){f.delete();}
+    File folder = new File(REP_TEMPORARY);
+    if(folder.isDirectory()){
+      File[] fichiers = folder.listFiles();
+      for (File f : fichiers ) {
+        if(isImage(f)){f.delete();}
+      }
     }
   }
   /**
